+++ topogram/nlp.py
@@ -1,7 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-
 import re
 import logging
 logger = logging.getLogger('topogram.nlp')
